requirements.txt
ultralytics
pillow
tensorflow
opencv-python
gtts
flask
flask-cors
numpy
